Small commit, mainly adding controllers
Added a homepage, projects and upload controller. Changed the links in the layout blade so they'd work, same in the projects.blade file. Added an upload blade file but haven't finished the form yet, will do that tomorrow.

Also fixed a small copy and paste mistake in the routes and fixed it so it links to the right controller.